refactor: improvements login module
