+++ src/login/login.service.ts
@@ -7,10 +7,11 @@ import {
 import { JwtService } from '@nestjs/jwt';
 import { UsersService } from '../users/users.service';
 import { IUsers } from '../users/interfaces/users.interface';
-import { JwtPayload } from './interfaces/jwt.payload';
+import { JwtPayload } from './interfaces/jwt-payload.interface';
 import { LoginDto } from './dto/login.dto';
 import { ConfigService } from '@nestjs/config';
 import { HashingService } from '../shared/hashing/hashing.service';
+import { ActiveUserData } from './interfaces/active-user-data.interface';
 
 @Injectable()
 export class LoginService {
@@ -21,14 +22,14 @@ export class LoginService {
     private readonly hashingService: HashingService,
   ) {}
 
-  private async validate(loginDto: LoginDto): Promise<IUsers> {
+  private async findUserByEmail(loginDto: LoginDto): Promise<IUsers> {
     return await this.usersService.findByEmail(loginDto.email);
   }
 
   public async login(
     loginDto: LoginDto,
   ): Promise<any | { status: number; message: string }> {
-    return this.validate(loginDto)
+    return this.findUserByEmail(loginDto)
       .then(async (userData) => {
         if (!userData) {
           throw new UnauthorizedException();
@@ -42,31 +43,33 @@ export class LoginService {
         if (!passwordIsValid == true) {
           return {
             message: 'Authentication failed. Wrong password',
-            status: 400,
+            status: HttpStatus.BAD_REQUEST,
           };
         }
 
-        const payload = {
+        return await this.signToken({
           name: userData.name,
           email: userData.email,
           id: userData.id,
-        };
-
-        const accessToken = this.jwtService.sign(payload);
-
-        return {
-          sub: payload.id,
-          expiresIn: this.configService.get('JWT_ACCESS_TOKEN_TTL'),
-          audience: this.configService.get('JWT_TOKEN_AUDIENCE'),
-          issuer: this.configService.get('JWT_TOKEN_ISSUER'),
-          accessToken: accessToken,
-          user: payload,
-          status: 200,
-        };
+        });
       })
       .catch((err) => {
         throw new HttpException(err, HttpStatus.BAD_REQUEST);
       });
+  }
+
+  private async signToken(payload: ActiveUserData) {
+    const accessToken = await this.jwtService.signAsync(payload);
+
+    return {
+      sub: payload.id,
+      expiresIn: this.configService.get('JWT_ACCESS_TOKEN_TTL'),
+      audience: this.configService.get('JWT_TOKEN_AUDIENCE'),
+      issuer: this.configService.get('JWT_TOKEN_ISSUER'),
+      accessToken: accessToken,
+      user: payload,
+      status: HttpStatus.OK,
+    };
   }
 
   public async validateUserByJwt(payload: JwtPayload) {
